+++ CHANGES.rst
@@ -3,10 +3,10 @@ Changelog
 
 Here you can see the full list of changes between each releases of ConsenSys-Utils.
 
-Version 0.2.0
--------------
+Version 0.2.0b1
+---------------
 
-Unreleased
+Released on August 6th 2018
 
 Feat